+++ src/main/java/net/ccbluex/liquidbounce/features/module/ModuleManager.kt
@@ -15,6 +15,7 @@ import net.ccbluex.liquidbounce.features.module.modules.`fun`.Derp
 import net.ccbluex.liquidbounce.features.module.modules.`fun`.SkinDerp
 import net.ccbluex.liquidbounce.features.module.modules.`fun`.SnakeGame
 import net.ccbluex.liquidbounce.features.module.modules.misc.*
+import net.ccbluex.liquidbounce.features.module.modules.misc.cheatdetector.CheatDetector
 import net.ccbluex.liquidbounce.features.module.modules.movement.*
 import net.ccbluex.liquidbounce.features.module.modules.movement.fly.Fly
 import net.ccbluex.liquidbounce.features.module.modules.movement.liquidwalk.LiquidWalk
@@ -92,6 +93,7 @@ object ModuleManager : Listenable, Collection<Module> by MODULE_REGISTRY {
             CameraView,
             Chams,
             Chat,
+            CheatDetector,
             ChestAura,
             ChestStealer,
             CivBreak,