+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/movement/fly/modes/spartan/Spartan.kt
@@ -10,11 +10,11 @@ import net.ccbluex.liquidbounce.utils.client.PacketUtils.sendPackets
 import net.ccbluex.liquidbounce.utils.extensions.component1
 import net.ccbluex.liquidbounce.utils.extensions.component2
 import net.ccbluex.liquidbounce.utils.extensions.component3
-import net.ccbluex.liquidbounce.utils.timing.TickTimer
+import net.ccbluex.liquidbounce.utils.timing.TickDelayTimer
 import net.minecraft.network.play.client.C03PacketPlayer.C04PacketPlayerPosition
 
 object Spartan : FlyMode("Spartan") {
-    private val timer = TickTimer()
+    private val timer = TickDelayTimer(12)
 
     override fun onEnable() {
         timer.reset()
@@ -24,15 +24,11 @@ object Spartan : FlyMode("Spartan") {
         mc.thePlayer?.apply {
             motionY = 0.0
 
-            timer.update()
-
-            if (timer.hasTimePassed(12)) {
+            if (timer.resetIfPassed()) {
                 sendPackets(
                     C04PacketPlayerPosition(posX, posY + 8, posZ, true),
                     C04PacketPlayerPosition(posX, posY - 8, posZ, true)
                 )
-
-                timer.reset()
             }
         }
     }