+++ src/main/java/net/ccbluex/liquidbounce/features/module/modules/movement/speed/modes/ncp/Frame.kt
@@ -25,11 +25,13 @@ object Frame : SpeedMode("Frame") {
                     tryJump()
                     if (motionTicks == 1) {
                         tickTimer.reset()
+
                         if (move) {
                             motionX = 0.0
                             motionZ = 0.0
                             move = false
                         }
+
                         motionTicks = 0
                     } else motionTicks = 1
                 } else if (!move && motionTicks == 1 && tickTimer.hasTimePassed(5)) {