+++ src/main/java/net/ccbluex/liquidbounce/utils/inventory/ArmorComparator.kt
@@ -116,7 +116,7 @@ class ArmorSet(private vararg val armorPairs: Pair<Int?, ItemStack>?) : Iterable
 
 		val baseDefense = baseDefensePercentage / 100f
 
-		// Not ceiling epf up to simulate the fact that 0.75f is actually random number between 0.5 and 1
+		// Not ceiling epf up to simulate the fact that 0.75f is actually a random number between 0.5 and 1
 		// By ceiling up, you for example get that 3x protection 1 is same as 4x protection 1, even tho 4x protection 1 has better overall average defense
 		// More details: https://www.guilded.gg/CCBlueX/groups/1dgpg8Jz/channels/034be45e-1b72-4d5a-bee7-d6ba52ba1657/chat?messageId=c0d88f1e-5ad6-48f3-8acb-d5ab7611164b
 		baseDefense + (1 - baseDefense) * epf.coerceAtMost(25) * 0.75f * 0.04f