+++ src/main/java/net/ccbluex/liquidbounce/utils/inventory/SilentHotbar.kt
@@ -89,9 +89,8 @@ object SilentHotbar : Listenable, MinecraftInstance {
     }
 
     fun renderSlot(option: Boolean): Int {
-        val player = mc.thePlayer ?: return 0
 
-        val original = player.inventory.currentItem
+        val original = mc.thePlayer.inventory.currentItem
 
         val state = hotbarState ?: return original
 
@@ -126,16 +125,14 @@ object SilentHotbar : Listenable, MinecraftInstance {
     }
 
     val onPacket = handler<PacketEvent> { event ->
-        val packet = event.packet
-
-        if (packet !is C09PacketHeldItemChange)
+        if (event.packet !is C09PacketHeldItemChange)
             return@handler
 
         /**
          * When a class (module usually) directly sends this packet and [SilentHotbarState.resetManually]
          * is active, it lets the class' slot change request override [currentSlot]
          */
-        if (shouldReset(packet.slotId, currentSlot))
+        if (shouldReset(event.packet.slotId, currentSlot))
             resetSlot()
     }